+++ server/src/test/passage-text-parsers/chapbook/chapbook-parser.test.ts
@@ -4010,69 +4010,6 @@ describe("Chapbook Parser", () => {
                 expect(result.range).to.eql(Range.create(1, 4, 1, 5));
             });
 
-            it("should error on a variable with a bad first character", () => {
-                const header = ":: Passage\n";
-                const passage = " 1var : 17\n--\n";
-                const callbacks = new MockCallbacks();
-                const state = buildParsingState({
-                    content: header + passage,
-                    callbacks: callbacks,
-                });
-                const parser = uut.getChapbookParser(undefined);
-
-                parser?.parsePassageText(passage, header.length, state);
-                const [result] = callbacks.errors;
-
-                expect(callbacks.errors.length).to.equal(1);
-                expect(result.severity).to.eql(DiagnosticSeverity.Error);
-                expect(result.message).to.include(
-                    "Variable names must start with a letter, $, or _"
-                );
-                expect(result.range).to.eql(Range.create(1, 1, 1, 2));
-            });
-
-            it("should error on a variable with illegal characters", () => {
-                const header = ":: Passage\n";
-                const passage = " väur😊 : 17\n--\n";
-                const callbacks = new MockCallbacks();
-                const state = buildParsingState({
-                    content: header + passage,
-                    callbacks: callbacks,
-                });
-                const parser = uut.getChapbookParser(undefined);
-
-                parser?.parsePassageText(passage, header.length, state);
-                const result = callbacks.errors;
-
-                expect(result.length).to.equal(2);
-                expect(result[0].severity).to.eql(DiagnosticSeverity.Error);
-                expect(result[0].message).to.include(
-                    "Must be a letter, digit, $, or _"
-                );
-                expect(result[0].range).to.eql(Range.create(1, 2, 1, 3));
-                expect(result[1].severity).to.eql(DiagnosticSeverity.Error);
-                expect(result[1].message).to.include(
-                    "Must be a letter, digit, $, or _"
-                );
-                // Note that this looks like 2 characters b/c of UTF-16 encoding of the smilie face
-                expect(result[1].range).to.eql(Range.create(1, 5, 1, 7));
-            });
-
-            it("should not error on a variable with a dot", () => {
-                const header = ":: Passage\n";
-                const passage = " var.sub : 17\n--\n";
-                const callbacks = new MockCallbacks();
-                const state = buildParsingState({
-                    content: header + passage,
-                    callbacks: callbacks,
-                });
-                const parser = uut.getChapbookParser(undefined);
-
-                parser?.parsePassageText(passage, header.length, state);
-
-                expect(callbacks.errors).to.be.empty;
-            });
-
             it("should error on an unclosed parenthesis before a colon", () => {
                 const header = ":: Passage\n";
                 const passage = "var(cond: 17\n--\n";